refactor: move the assignment document into its own module [REPORT-81]
Create assignment-doc.ts holding computeAssignmentDocId and
getAlternatingAssignment, moved verbatim, and import them into
random-assignment.ts. Their tests move to assignment-doc.test.ts.

The module imports nothing but crypto and firebase-admin, which is what
lets a later emulator test import real production code with no mocks:
firebase-functions pulls in firebase-admin/auth (jest 24 cannot resolve
its subpath exports) and firebase-client pulls in firebase/auth (fetch
is not defined under this jest environment). A header comment records
the constraint.

No behaviour change: 9 suites, 224 tests, lint and tsc clean.
