+++ functions/src/tasks/ai4vs-flvs/random-assignment.test.ts
@@ -60,7 +60,7 @@ jest.mock("firebase-admin", () => ({
 }));
 
 // Must import after jest.mock
-import { randomAssignment, computeAssignmentDocId, getAlternatingAssignment } from "./random-assignment";
+import { randomAssignment } from "./random-assignment";
 
 /** Build a mock Firestore answer doc with the given prompt and selected choices. */
 const makeAnswerDoc = (
@@ -820,155 +820,5 @@ describe("randomAssignment", () => {
         expect.stringContaining("interactiveId")
       );
     });
-  });
-});
-
-describe("computeAssignmentDocId", () => {
-  it("produces a deterministic hex string for the same inputs", () => {
-    const id1 = computeAssignmentDocId("int-1", "https://learn.concord.org", "rl-1", "ctx-1");
-    const id2 = computeAssignmentDocId("int-1", "https://learn.concord.org", "rl-1", "ctx-1");
-
-    expect(id1).toBe(id2);
-    expect(id1).toMatch(/^[a-f0-9]{64}$/);
-  });
-
-  it("produces different IDs for different inputs", () => {
-    const id1 = computeAssignmentDocId("int-1", "https://learn.concord.org", "rl-1", "ctx-1");
-    const id2 = computeAssignmentDocId("int-2", "https://learn.concord.org", "rl-1", "ctx-1");
-    const id3 = computeAssignmentDocId("int-1", "https://learn.concord.org", "rl-1", "ctx-2");
-
-    expect(id1).not.toBe(id2);
-    expect(id1).not.toBe(id3);
-    expect(id2).not.toBe(id3);
-  });
-});
-
-describe("getAlternatingAssignment", () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-    mockTransactionGet.mockResolvedValue({ exists: false, data: () => undefined });
-  });
-
-  it("returns n1 assignment when document does not exist (first student)", async () => {
-    const result = await getAlternatingAssignment(
-      "src-1", "int-1", "plat-1", "rl-1", "ctx-1", "user-1", "Female|White|High|Mod1", "control"
-    );
-
-    expect(result).toBe("control");
-  });
-
-  it("returns opposite of n1 when nextAssignment is set (second student)", async () => {
-    mockTransactionGet.mockResolvedValue({
-      exists: true,
-      data: () => ({
-        strata: {
-          "Female|White|High|Mod1": {
-            nextAssignment: "treatment",
-            users: { "user-1": "control" },
-          },
-        },
-      }),
-    });
-
-    const result = await getAlternatingAssignment(
-      "src-1", "int-1", "plat-1", "rl-1", "ctx-1", "user-2", "Female|White|High|Mod1", "control"
-    );
-
-    expect(result).toBe("treatment");
-  });
-
-  it("returns n1 again on third student (verifies continued alternation)", async () => {
-    mockTransactionGet.mockResolvedValue({
-      exists: true,
-      data: () => ({
-        strata: {
-          "Female|White|High|Mod1": {
-            nextAssignment: "control",
-            users: { "user-1": "control", "user-2": "treatment" },
-          },
-        },
-      }),
-    });
-
-    const result = await getAlternatingAssignment(
-      "src-1", "int-1", "plat-1", "rl-1", "ctx-1", "user-3", "Female|White|High|Mod1", "control"
-    );
-
-    expect(result).toBe("control");
-  });
-
-  it("returns cached assignment without flipping nextAssignment on dedup", async () => {
-    mockTransactionGet.mockResolvedValue({
-      exists: true,
-      data: () => ({
-        strata: {
-          "Female|White|High|Mod1": {
-            nextAssignment: "treatment",
-            users: { "user-1": "control" },
-          },
-        },
-      }),
-    });
-
-    const result = await getAlternatingAssignment(
-      "src-1", "int-1", "plat-1", "rl-1", "ctx-1", "user-1", "Female|White|High|Mod1", "control"
-    );
-
-    expect(result).toBe("control");
-    // Should NOT have written to the document
-    expect(mockTransactionSet).not.toHaveBeenCalled();
-  });
-
-  it("creates document with correct structure on first call", async () => {
-    await getAlternatingAssignment(
-      "src-1", "int-1", "plat-1", "rl-1", "ctx-1", "user-1", "Female|White|High|Mod1", "control"
-    );
-
-    expect(mockTransactionSet).toHaveBeenCalledWith(
-      mockDocRef,
-      expect.objectContaining({
-        type: "ai4vs-flvs-assignments",
-        interactiveId: "int-1",
-        platform_id: "plat-1",
-        resource_link_id: "rl-1",
-        context_id: "ctx-1",
-        strata: {
-          "Female|White|High|Mod1": {
-            nextAssignment: "treatment",
-            users: { "user-1": "control" },
-          },
-        },
-      }),
-      { merge: true }
-    );
-  });
-
-  it("flips nextAssignment to opposite after each new assignment", async () => {
-    // First student gets "treatment" (n1), nextAssignment should flip to "control"
-    await getAlternatingAssignment(
-      "src-1", "int-1", "plat-1", "rl-1", "ctx-1", "user-1", "Female|White|High|Mod2", "treatment"
-    );
-
-    expect(mockTransactionSet).toHaveBeenCalledWith(
-      mockDocRef,
-      expect.objectContaining({
-        strata: expect.objectContaining({
-          "Female|White|High|Mod2": expect.objectContaining({
-            nextAssignment: "control",
-          }),
-        }),
-      }),
-      { merge: true }
-    );
-  });
-
-  it("throws on transaction failure", async () => {
-    mockRunTransaction.mockRejectedValueOnce(new Error("transaction failed"));
-
-    await expect(
-      getAlternatingAssignment(
-        "src-1", "int-1", "plat-1", "rl-1", "ctx-1", "user-1", "Female|White|High|Mod1", "control"
-      )
-    ).rejects.toThrow("transaction failed");
   });
 });
